# use strings.Split split subject