This PR is use strings.Split to split msg subject in sublist.go