@kozlovic 2017-08-15
Thank you for your interest!

Unfortunately, we are not going to merge this PR. You can run a benchmark between the original approach and the strings.Split() approach and you will see why we did the way we did. The strings.Split() will be much slower and double the amount of memory allocation on a given Insert, etc.. (especially when there are more than one token).

Here is an example of trying to `Insert()` the same subscription with a subject of 5 tokens (`foo.foo.foo.foo.foo`) with master (`old`) and your branch (`new`):
```
Ivans-MacBook-Pro:gnatsd ivan$ benchcmp old.txt new.txt 
benchmark                                         old ns/op     new ns/op     delta
Benchmark_______SublistInsertMultipleTokens-8     330           613           +85.76%

benchmark                                         old allocs     new allocs     delta
Benchmark_______SublistInsertMultipleTokens-8     1              2              +100.00%

benchmark                                         old bytes     new bytes     delta
Benchmark_______SublistInsertMultipleTokens-8     75            158           +110.67%
```
In this case the delta being positive is not a good thing since it means that it takes that much longer (or that much more memory) for individual inserts.